requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
asttokens==3.0.0
boto3==1.39.0
botocore==1.39.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.0.0
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
holidays==0.75
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-core==0.3.67
langchain-text-splitters==0.3.8
langsmith==0.4.4
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.1
orjson==3.10.18
packaging==24.2
pandas==2.3.0
parso==0.8.4
patsy==1.0.1
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
prophet==1.1.7
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
requests-toolbelt==1.0.0
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.33.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.4
tenacity==9.1.2
threadpoolctl==3.6.0
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
zstandard==0.23.0